little refactoring of tools
