--- lab_4/consts.py
@@ -1,2 +1,0 @@
-BINS_FILE = 'files/bins.json'
-CARD_NUMBER_FILE = 'files/card_num.json'
