+++ lab_4/tools/card_hash.py
