+++ lab_4/tools/consts.py
@@ -1,0 +1,2 @@
+BINS_FILE = '../files/bins.json'
+CARD_NUMBER_FILE = '../files/card_num.json'
